==> notebooks/1-data_preprocessing.ipynb <==
import os
import pandas as pd
import numpy as np
import sys
import warnings
warnings.filterwarnings("ignore")

# PROJECT_ROOT = os.getcwd()  
# sys.path.append(os.path.join(PROJECT_ROOT))
RAW_DATA_PATH = "../data/data_raw"
EXTRACT_DIR = "../data/data_extracted"
 
sys.path.append(os.path.join('../src'))

from data_loader import DataLoader

data_loader = DataLoader(RAW_DATA_PATH)
df = data_loader.load_json_data()

# to save the raw extracted data
# path_to_save = os.path.join(EXTRACT_DIR, 'raw_extracted_data.csv')
# df.to_csv(path_to_save, index=False)
# print('SUCCESS')

### Alternative Import
import it from saved ones



# df = pd.read_csv('../data/data_extracted/raw_extracted_data.csv', index_col=0, parse_dates=["time"])
df = pd.read_csv('../data/data_extracted/raw_extracted_data.csv', parse_dates=["time"])
df.head()

df.info()

df.isna().sum()

# missing values
df["time"] = df["time"].fillna(method="ffill")

# sort
df.set_index('time', inplace=True)
df = df.sort_index()

df['time_diff'] = df.index.to_series().diff().dt.total_seconds()
# df['time_diff'].describe()

from scipy.signal import butter, filtfilt
fs = 1 / df['time_diff'].median()  # Hz (sampling frequency)

def low_pass_filter(data, cutoff=5, fs=fs, order=4):
    nyquist = 0.5 * fs  # Nyquist Frequency
    normal_cutoff = cutoff / nyquist
    b, a = butter(order, normal_cutoff, btype='low', analog=False)
    return filtfilt(b, a, data)


# only applying filter on gyroscope because there are a lot of fluctuation 
sensor_columns = ["gx", "gy", "gz"]
# sensor_columns = ["ax", "ay", "az", "gx", "gy", "gz"]

for col in sensor_columns:
    df[col] = low_pass_filter(df[col])


df['acc_magnitude'] = np.sqrt(df['ax']**2 + df['ay']**2 + df['az']**2)
df['gyro_magnitude'] = np.sqrt(df['gx']**2 + df['gy']**2 + df['gz']**2)

### Feature Engineering

# Reset index to bring 'time' back as a column
df_reset = df.reset_index()

# Group by 'id' and aggregate numerical features
session_features = df_reset.groupby("id").agg({
    "ax": ["mean", "std", "min", "max"],
    "ay": ["mean", "std", "min", "max"],
    "az": ["mean", "std", "min", "max"],
    "gx": ["mean", "std", "min", "max"],
    "gy": ["mean", "std", "min", "max"],
    "gz": ["mean", "std", "min", "max"],
    "acc_magnitude": ["mean", "std", "min", "max"],
    "gyro_magnitude": ["mean", "std", "min", "max"],
}).reset_index()

# Flatten MultiIndex column names
# session_features.columns = ["_".join(col).strip() if isinstance(col, tuple) else col for col in session_features.columns]
session_features.columns = ["_".join(col).strip() for col in session_features.columns.values]
session_features = session_features.rename(columns={"id_": "id"})

# Compute start_time and end_time
time_stats = df_reset.groupby("id")["time"].agg(["min", "max"]).reset_index()
time_stats.rename(columns={"min": "start_time", "max": "end_time"}, inplace=True)

# Calculate session duration in seconds
time_stats["session_duration"] = (time_stats["end_time"] - time_stats["start_time"]).dt.total_seconds()

# Calculate num_measurements using size()
num_measurements = df_reset.groupby("id").size().reset_index(name="num_measurements")

# Merge all computed features
session_features = session_features.merge(time_stats, on="id").merge(num_measurements, on="id")

# in order to see all the columns in dataframe
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

session_features.head()

### Calculating the Height

# using mean of acc_magnitude
mean_acc_magnitude = df["acc_magnitude"].mean()
std_acc_magnitude = df["acc_magnitude"].std()

mean_height = mean_acc_magnitude + 0.5 * std_acc_magnitude  

peaks, _ = find_peaks(df["acc_magnitude"], height=mean_height)

print("peaks", len(peaks))
print("mean_az",mean_acc_magnitude)
print("std_az",std_acc_magnitude)
print("height", mean_height)

### Finding the steps

from scipy.signal import find_peaks

fs = 1 / df['time_diff'].median()  # Sampling rate in Hz (samples / seconds)
min_distance = int(0.5 * fs)  # Minimum distance between peaks (0.5 seconds)
# every step takes around 0.5 seconds

def count_peaks(series):
    peaks, _ = find_peaks(series, height=mean_height, distance=min_distance)  
    return len(peaks)

# Count steps based on side (L or R)
step_counts = df.groupby(["id", "side"])["acc_magnitude"].apply(count_peaks).unstack(fill_value=0)

# Check if 'side' values contain 'L' (left) and 'R' (right), then assign them explicitly
step_counts['left_steps'] = step_counts.get('L', 0)
step_counts['right_steps'] = step_counts.get('R', 0)

# Remove the temporary 'L' and 'R' columns to avoid keeping them in the final output
step_counts = step_counts.drop(columns=['L', 'R'], errors='ignore')

session_features = session_features.rename(columns={"id_": "id"})
final_df = session_features.merge(step_counts, on="id")

final_df.head()


display(final_df[['left_steps', 'right_steps', 'id', 'acc_magnitude_mean', 'session_duration', 'num_measurements']].max())
print("*" * 60)
display(final_df[['left_steps', 'right_steps', 'id', 'acc_magnitude_mean', 'session_duration', 'num_measurements']].min())

### Saving Preprocessed Data:

# to save the raw extracted data
path_to_save = os.path.join(EXTRACT_DIR, 'preprocessed_data.csv')
final_df.to_csv(path_to_save, index=False)
print('SUCCESS')

from watermark import watermark
print(watermark())
