==> notebooks/3-pipeline.ipynb <==
## Pipeline
streamlining some real world scenario of operationalising the model. <br><br>

- step between time A and time B 
- step counts from point A till 15 minutes or 20 minutes or more

Important Note: <br>
select the time in the interval of following: <br>
the entire dataset spans from 07:19:32.380 to 13:03:10.196 on June 14, 2024, which is approximately 5 hours and 43 minutes (~20,218 seconds)

import os
import pandas as pd
import numpy as np
import sys
import warnings
warnings.filterwarnings("ignore")

# paths
RAW_DATA_PATH = "../data/data_raw"
EXTRACT_DIR = "../data/data_extracted"
OUTPUT_DIR = "../data/data_output"
MODEL_PATH = "../models/random_forest_model.pkl"

 
sys.path.append(os.path.join('../src'))

from data_loader import DataLoader
from data_processor import DataPreprocessor, FeatureExtractor
from steps_predictor import StepPredictor, StepPredictionPipeline


### Running Pipeline
you can run the following pipeline from project root as well. <br>
in project root:
```
python src/steps_predictor.py
```
make sure you have installed packages in requirements.txt and have the have sensor data in data_raw


import json
from datetime import datetime

print('Processing - please wait ...')
data_loader = DataLoader(RAW_DATA_PATH)
preprocessor = DataPreprocessor()
predictor = StepPredictor(MODEL_PATH)
pipeline = StepPredictionPipeline(data_loader, preprocessor, predictor)
prediction_result = pipeline.run()

# Save predictions to a unique JSON file
timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
output_file = os.path.join(OUTPUT_DIR, f"predictions_{timestamp}.json")
with open(output_file, "w") as f:
     json.dump(prediction_result, f, indent=4)
     
print("SUCCESS")

from watermark import watermark
print(watermark())
